Add writable integer storage with source-owned target proofs (#222)
Accept #221 after independent review, repro-driven fixes and recheck plus parent typed-family/store-conversion/identity audit. Corrected capture-write and source-empty-function proof gaps, including forged definition/transfer metadata. Source046c432cffbef0dcbb9b9f0b15bf7a24e2655883/tree45fd1db24b7c4c65ef62a52a30c79e23a187e585 exactly combines reviewed fixes with accepted InitLocals. Frozen462 focused/973 full per configuration and four owning-assembly all-stage captures passed; generated hash manifests verified. Compute-only f32/i32/u32, exact resultless stores, no new synchronization guarantees. C native bit/control/reset acceptance remains separate and now unblocked. Topic only, no main merge.

Copilot-Session: f8d9fa08-f0d8-4c5e-90ce-734c41b3599f

+++ DualDrill.CLSL.Test/ReadonlyStructuredBufferTests.cs
@@ -190,7 +190,7 @@ public sealed class ReadonlyStructuredBufferTests(ITestOutputHelper output)
         Assert.Throws<NotSupportedException>(() =>
             new RuntimeReflectionParser().ParseType(typeof(StructuredBuffer<double>)));
         Assert.Throws<NotSupportedException>(() =>
-            new RuntimeReflectionParser().ParseType(typeof(RWStructuredBuffer<int>)));
+            new RuntimeReflectionParser().ParseType(typeof(RWStructuredBuffer<double>)));
         Assert.NotNull(StructuredBufferLoadOperation<FloatType<N32>>.Instance);
     }
